requirements.txt
sentence-transformers
numpy
pandas
scikit-learn
google-generativeai
faiss-cpu
streamlit
pytest
plotly